Stop reading PNGs after the IEND chunk
If there are extra bytes at the end of the PNG (e.g. if the PNG is
embedded in some other sort of file), the parser will try to read in new
chunks even if they're nonsensical, and will often fail because of
invalid data in the garbage chunk (like a "chunk size" larger than the
file) By stopping at the IEND chunk, the parser won't keep trying to
read more chunks past the end of the file.

According to [5.6 Chunk Ordering](https://www.w3.org/TR/PNG/#5ChunkOrdering)
in the PNG spec, the (only) IEND chunk is the last chunk in the file,
and the IEND chunk specification says that it "marks the end of the PNG
datastream": [11.2.5 IEND Image Trailer](https://www.w3.org/TR/PNG/#11IEND)

Looking at libpng, `png_read_end` stops reading as soon as the
`PNG_HAVE_IEND` mode flag is set on the PNG, and sets that flag
(via `png_handle_iend`) as soon as it sees the IEND chunk.