TrueType 'post' goes to eos.
The current implementation assumes that there will be a zero length
pascal string at the end of a version 2 'post' table. There is no such
requirement and this will fail to load a font which happens to have such
a table which ends on exactly a four byte boundary.

Note that FreeType determines the number of pascal strings to read by
iterating over the glyph name index values and selecting the highest
index. However, this only shows used strings and not all of the data.