Replace member_name type switch with separate conditional instances
This is a workaround. Only one of the three instances is present at
any time, so a type switch would be more appropriate. However, when an
instance's type is switched, it is impossible to refer to the fields
of any specific type case (since it is not statically known if that
case is actually the one that was chosen).

By using three separate instances, we can refer to the fields of any
of the instances, but have to manually check that the instance in
question is actually present.