Add convenience instances for ar space_padded_number fields
The main advantage is that this makes the specs easier to use - most
users only care about the parsed value, so it should be easily
accessilbe without an extra .value each time. Those who need the
unparsed text value can access it via the ..._raw fields.

This way the four ar variants now also have the same interface.
For example, a member's data size is now always accessed as
member.size, where previously it was sometimes member.size (ar_bsd) and
sometimes member.size.value (all other specs). This allows writing
generic code that can work with any of the ar specs.

Currently the compiler knows nothing about this common interface, which
means that using the interface generically is only possible from
dynamically-typed languages, such as Python. If something similar to
kaitai-io/kaitai_struct#314 is implemented in the future, the interface
could be properly specified in Kaitai Struct so that the compiler can
understand, enforce and properly expose it.