Add Android super.img (Dynamic Partitions) format (#349)
The metadata stored by Android at the beginning of a "super" partition,
which is what it calls a disk partition that holds one or more Dynamic
Partitions. Dynamic Partitions do more or less the same thing that LVM
does on Linux, allowing Android to map ranges of non-contiguous extents
to a single logical device. This metadata holds that mapping.

Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>