common/tzif.ksy: Add _flags suffix to is_ut and is_std
The `is_ut` and `is_std` members of tz_data are arrays, so their names
should be pluralized.  However, the names `is_uts` and `is_stds` are
kind of awkward.  We already have the corresponding `num_is_ut_flags`
and `num_is_std_flags`, so change the names of the arrays to
`is_ut_flags` and `is_std_flags` to match.