elf: pull padding from `note_section_entry` fields
See https://docs.oracle.com/cd/E19683-01/816-1386/6m7qcoblj/index.html#chapter6-18048:

"Padding is present, if necessary, to ensure 4-byte alignment for the descriptor. Such padding is not included in `namesz`."

"Padding is present, if necessary, to ensure 4-byte alignment for the next note entry. Such padding is not included in `descsz`."