Add Amlogic eMMC partition format, as used by Meson SoCs
This is an unnamed and undocumented partition table format implemented
by the bootloader and kernel that Amlogic provides for their Linux SoCs
(Meson series at least, and probably others). They appear to use this
rather than GPT, the industry standard, because their BootROM loads and
executes the next stage loader from offset 512 (0x200) on the eMMC,
which is exactly where the GPT header would have to start. So instead of
changing their BootROM, Amlogic devised this partition table, which
lives at an offset of 36MiB (0x240_0000) on the eMMC and so doesn't
conflict. This parser expects as input just the partition table from
that offset. The maximum number of partitions in a table is 32, which
corresponds to a maximum table size of 1304 bytes (0x518).

Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>