mach_o: rename entitlement type, fix constant name
We currently call this "entitlement", but for parity with the source it
should really be "entitlements" (or "embedded_entitlements", but that's
quite long and currently an unneeded distinguisher). XNU's constant for
this is CSMAGIC_EMBEDDED_ENTITLEMENTS[1], not CSMAGIC_ENTITLEMENT like
our comment claims. Change the type name and fix the comment as well.

[1] https://opensource.apple.com/source/xnu/xnu-7195.121.3/osfmk/kern/cs_blobs.h.auto.html