mach_o: parse all binding opcodes instead of stopping at done
Apple's dyld has two end conditions for parsing: a REBASE_OPCODE_DONE
opcode or the end of the section (see newly added doc-ref). Currently,
the Kaitai parser only looks for the opcode and will fail if the section
ends without one.

I have encountered a binary with no done opcode, so some toolchains
clearly choose to omit it. Given this, it seems more reliable to parse
the whole section like we do with lazy_bind and weak_bind already.
Linkers shouldn't insert a done opcode before the end of the section
and, if they do, I think it's reasonable for Kaitai to show that instead
of pretending that the section ends at the first done opcode.