elf: fix excluding range endpoints for tag encoding rules
See
https://docs.oracle.com/cd/E37838_01/html/E36783/chapter6-42444.html#OSLLGchapter6-tbl-52.
But beware of occasional inaccurate information there (enclosed in
`~~`):

> DT_ENCODING > Dynamic tag values that are greater than or equal to
DT_ENCODING, and less than ~~or equal to~~ DT_LOOS, follow the rules for
the interpretation of the d_un union.

The value of `DT_LOOS` is `0x6000000d`, but `0x6000000d` is also
assigned to `DT_SUNW_AUXILIARY`, which uses the `d_un.d_ptr` field of
the dynamic structure. However, `0x6000000d` is an odd number (`0xd` is
13), which would indicate that a tag assigned to this value (i.e.
`DT_SUNW_AUXILIARY`) uses d_val. So `DT_SUNW_AUXILIARY` certainly **does
not** follow the common `d_un` interpretation, and thus the value of
`DT_LOOS` must be **excluded** from the range of tags that follow the
interpretation (so this was already correct in the elf.ksy spec, no fix
was necessary here).