pif: check `len_file >= ofs_image_data_min` explicitly
Although this was already implicitly enforced by the combination of
`valid/{min,max}` in `ofs_image_data`, it's better to do it directly in
`len_file` too. The reason is that if it wasn't true, the validation
error would now only be reported in `ofs_image_data`, but in fact there
would be no valid value for `ofs_image_data` because `len_file` was
already wrong.