vlq_base128_le: refactor for full 64-bit support
More precisely, 64-bit precision should be now available in all
applicable languages (in KSC 0.11, Perl is not applicable because
`params` are not implemented, Go is not applicable because type casting
`.as<>` is not implemented, and Rust is not applicable because for some
reason it does `(... as i32)` everywhere).

In JavaScript, the 64-bit support is now best-effort (previously, all
results were truncated to 32 bits). This means that if you end up with
[unsafe integers](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/isSafeInteger#description),
you will obviously lose some precision in the least significant bits,
but it will still be quite close to the correct result (I think the
result may be off by ~1024 at most, but don't quote me on that).

Working with 64-bit integers in Kaitai Struct so that it works
consistently across languages shouldn't be this difficult. We should
implement proper support for 64-bit integers in JavaScript
(https://github.com/kaitai-io/kaitai_struct/issues/183) because that's
basically the only thing that can't be worked around. But while working
on this commit, I encountered some annoying problems in other languages
(PHP, C#) too, as explained in the code comments.

This commit was inspired by the following note in
<https://github.com/mitmproxy/mitmproxy/pull/4872#issue-1030312029>:

> Also, the Kaitai based parser has been replaced with a built-in
> parser, because it did not follow the specs (the base128 encoding of
> the Kaitai decoder only respects variable length integers up to 56bit,
> while the protobuf specs support variable length integers up to
> 64bit).