sqlite3: remove incorrect type casting
The `column_content` type's `ser` parameter was declared as `struct`
(i.e. an arbitrary user-defined type) and unconditionally casted to
`serial` using `.as<serial>`. However, `column_content` was actually
always invoked with a `vlq_base128_be` object being passed as a
parameter instead of `serial`, which would (or at least *should*) fail
in any statically-typed language.

On the first glance, this mistake is related to
https://github.com/kaitai-io/kaitai_struct/issues/696. Unfortunately,
this specific case probably wouldn't be a compile error even when
https://github.com/kaitai-io/kaitai_struct/issues/696 is implemented,
because individually, both steps "1. passing `vlq_base128_be` into a
`struct` parameter" and "2. downcasting a `struct` parameter to
`serial`" appear to be valid.