tcp_segment: move `to-string` between `doc` and `seq`
This matches the place where I put it in ksy_schema (see
https://github.com/kaitai-io/ksy_schema/commit/1bd49f16f438c1995e892201f4b8def38ff51c7d).
I'd argue it makes more sense at the top of the type definition than at
the bottom. `to-string` is a successor of `-webide-representation`,
whose purpose is to show a quick summary (i.e. short string
representation) of the object in visualizers even before expanding it to
reveal its fields. The type definition can be long, and it doesn't seem
reasonable to have to scroll all the way to the end just to see the
definition of this brief string representation, which is displayed
prominently in visualizers.